Get identifiers for neurons/lines with NeuronBridge results

This function gets all of the neuron identifiers or line codes (e.g.R24A02 is a generation 1 GMR GLA4 line, LH173 is a split-GAL4 line, 5813055865 is a hemibrain neuron) for published neurons/lines for which NeuronBridge has precomputed matching results. If you query is not among these identifiers you will not be able to get results for it.

neuronbridge_ids(version = "v2_1_1")

Arguments

version
the precomputed scores to search. For example, "v2_1_1" refers to this release.

Value

a vector of identifiers

See also

Examples

# \donttest{
all.nb.ids = neuronbridge_ids()
length(all.nb.ids)

#> [1] 36003

# }