Get information on neurons and lines used by NeuronBridge

This function gets information stored on NeuronBridge for *D. melanogaster* neurons and genetic driver lines that have been used in EM-LM searches.

neuronbridge_info(
  id,
  dataset = c("detect", "by_line", "by_body"),
  version = "v2_1_1",
  ...
)

Arguments

id

character vector. An identifier/identifiers for the neuron(s)/line(s) you wish to search. This can either be a line code for a GAL4 line (e.g. R16F12) or a split GAL4 line (e.g. LH173) or a 'body ID' for a neuron from the hemibrain connectome, (e.g. 1537331894).

dataset

whether the ID given is a body ID ("by_body") from the hemibrain connectome or a genetic driver line ("by_line"). If left at "detect" then neuronbridger tries to guess to which id belongs by using neuronbridger:::guess_dataset.

version

the precomputed scores to search. For example, "v2_1_1" refers to this release.

...

methods passed to neuronbridger:::neuronbridge_fetch.

Value

a data.frame where each row corresponds to a MIP file that has been used to compute matching scores. Each relates to the id you gave in the function call, but each also has its own unique nb.id.

"publishedName"

- the id you gave in the function call.

"libraryName"

- the data set from which this data item came.

"imageURL"

- the path on https://s3.amazonaws.com/, at which one can find a 'high res' .png of the MIP file for this data item.

"thumbnailURL"

- the path on https://s3.amazonaws.com/, at which one can find a 'low res' .jpg thumbnail of the MIP file for this data item.

"gender"

- the sex of the fly brain which this data item derives

"nb.id"

- the 'NeuronBridge ID' for the MIP file.

"no.images"

- the total number of MIP files (each with its own nb.id) that related to the id for which you searched.

See also

Examples

if (FALSE) { # \dontrun{
# Get information on a mushroombody related GMR GAL4 driver line
nb.info.lm = neuronbridge_info("MB543B")

# Get information on the hemibrain neuron,
nb.info.em = neuronbridge_info("542634818")
} # }