Use NeuronBridge to estimate which hemibrain connectome neurons are taregted by a genetic driver line

This function uses neuronbridge_search to get potential connectime matches for the given genetic driver line (line). Each line may have multiple images associated with it, likely because it has undergone stochastic labelling to better see what the line contains. Each of these images is searched and potential connectome hits returned. Each is assigned its highest score. The final output is filtered to that the LM-EM matching score is below the given threshold.

neuronbridge_line_contents(
  line,
  neuprintr = TRUE,
  version = "v2_1_1",
  threshold = 23000
)

neuronbridge_predict_split(
  line1,
  line2,
  neuprintr = TRUE,
  version = "v2_1_1",
  threshold1 = 23000,
  threshold2 = 23000
)

Arguments

line

a single ID for a genetic driver line (in neuronbridge_ids) for which you want to find hits.

neuprintr

logical, whether or not to use the package neuprintr to fetch meta data (e.g. cell body fiber tract, cell type) on the connectome neurons returned. If TRUE then neuprintr::neuprint_get_meta is used, and additional columns from this call are added to the returned data.frame.

version

the precomputed scores to search. For example, "v2_1_1" refers to this release.

threshold

LM-EM matches with a normalizedScore below this value are not returned.

line1

a line you want to try combining with line2. The aim is to see which neurons are in both line1 and line2

line2

a line you want to try combining with line1. The aim is to see which neurons are in both line1 and line2

threshold1

LM-EM matches with a normalizedScore below this value are not returned. Applied to 'hits' for line1.

threshold2

LM-EM matches with a normalizedScore below this value are not returned. Applied to 'hits' for line2.

Value

a data.frame of hits. Each row idnciates a separate MIP file with its own nb.id. The data.frame is already ranked by normalizedScore. Top scores (better match) are at the top of the data frame. The columns mean:

"publishedName"

- the id for the potential hit neuron/line. I.e. specififes a genetic driver resource or a connectome neuron. these are the same ids that can be seen with neuronbridge_ids.

"libraryName"

- the data set from which this data item came.

#'
"normalizedScore"

- the matching score, created by examining the overlapped pixel number and colourdepth. If the colourand xy position of the pixel match between the mask and the searching data, then the approach here will count it as a positive matching score

"searched"

- the line you searched with, i.e. given to the function call

See also

Examples

if (FALSE) { # \dontrun{
# Interesting line that labels mushroom body neurons
## But which ones?
line = "MB543B"

# Let's find out
contents = neuronbridge_line_contents(line = line, threshold = 23000)
## Note! Choosing the right threshold for the nomarlsied MIP-comparison score
### Is critical. This may take some titring for lines you are really interested in
#### Though this value is a good first pass.

# So this is what is likely in this line!
## Note we have meta-data from neuprint! So we can see neuron types and names!!
# View(contents)

# Now let's check what these neurons look like
if (!require("neuprintr")) remotes::install_github("natverse/neuprintr")

# Load package
library(neuprintr)
## Note you need to 'login' to neuPrint through R
### Look at the package README and/or examine: ?neuprint_login

# Let's see the EM neurons all together
neurons = neuprintr::neuprint_read_neurons(unique(contents$publishedName))
plot3d(neurons)

# And compare with the LM data:
open3d()
mips = neuronbridge_mip(line)
scan_mip(mips,type="images", sleep = 5)

} # }