Make a .nrrd file from neurons in a given flywire scene

Creates a common image file (a .nrrd file) readable in Fiji from a flywire scene. User must have access to the flywire.ai production data set. See R package fafbseg scene for more information on accessing flywire data from R.

ngl_scene_to_nrrd(
  scene,
  reference = "JRC2018U_HR",
  savefolder = file.path(getwd(), "nrrd"),
  file = NULL
)

root_id_to_nrrd(
  root_ids,
  reference = "JRC2018U_HR",
  savefolder = file.path(getwd(), "nrrd"),
  file = NULL
)

Arguments

scene

a neuroglancer link to a flywire scene. You can get this by hitting the 'share' button in the flywire web browser GUI.

reference

the templatebrain to transform the flywire neuron(s) into before creating the .nrrd file. Defaults to "JRC2018U_HR", which is the brain space into which a lot of data for neuronbridge is formatted.

savefolder

a folder in which to save .nrrd files. defaults to a folder called 'nrrd' that it will create in your working directory.

file

the name of your .nrrd file. Defaults to the root IDs for the given flywire neurons, concatenated with the name of the reference brain space.

root_ids

a vector of root IDs that specify flywire meshes.

Value

a .nrrd file.

See also

Examples

# \donttest{
# root_id_to_nrrd(root_ids = c("720575940632668823", "720575940644278051"))
# ascending neurons,
# if this link does not work for you,
# you do not yet have access to the data set
# }