Get EM-LM neuron matches for fly brain neurons from neuronbridge.janelia.org

The function neuronbridge_hits retreives a ranked list of hits for a MIP file search. Specify the MIP file you want to search for by using its nb.id. This can be found by using neuronbridge_info. neuronbridge_search will run neuronbridge_info and then neuronbridge_hits, to make things easier. The function neuronbridge_search can also handle being given multiple IDs, i.e. searching for multiple data items at once. MIP options can be fetched and visualised with neuronbridge_mip.

neuronbridge_hits(nb.id, version = "v2_1_1")

neuronbridge_search(
  id,
  version = "v2_1_1",
  dataset = c("detect", "by_line", "by_body"),
  threshold = 10000
)

Arguments

nb.id
an internal ID used by neuronbridge.janelia.org to identigy neurons/lines. This differs from id used by other functions, and can be found using neuronbridge_info.

version
the precomputed scores to search. For example, "v2_1_1" refers to this release.

id
character vector. An identifier/identifiers for the neuron(s)/line(s) you wish to search. This can either be a line code for a GAL4 line (e.g. R16F12) or a split GAL4 line (e.g. LH173) or a 'body ID' for a neuron from the hemibrain connectome, (e.g. 1537331894).

dataset
whether the ID given is a body ID ("by_body") from the hemibrain connectome or a genetic driver line ("by_line"). If left at "detect" then neuronbridger tries to guess to which id belongs by using neuronbridger:::guess_dataset.

threshold
LM-EM matches with a normalizedScore below this value are not returned. If set to NULL, the results are not filtered.

Value

a data.frame of hits. Each row idnciates a separate MIP file with its own nb.id. The data.frame is already ranked by normalizedScore. Top scores (better match) are at the top of the data frame. The columns mean:

"publishedName" - the id for the potential hit neuron/line. I.e. specififes a genetic driver resource or a connectome neuron. these are the same ids that can be seen with neuronbridge_ids.

"libraryName" - the data set from which this data item came.

"imageURL" - the path on https://s3.amazonaws.com/, at which one can find a 'high res' .png of the MIP file for this data item.

"thumbnailURL" - the path on https://s3.amazonaws.com/, at which one can find a 'low res' .jpg thumbnail of the MIP file for this data item.

"slideCode" - the unique identifier for the sample from which the MIP came. The first number indicates the date the image was taken by FlyLight.

"objective" - the magnification under which the imasge was taken.

"gender" - the sex of the fly brain which this data item derives. f = female, m = male.

"anatomicalArea" - the gross part of the nervous system images, e.g. brain or ventral nervous system.

"alignmentSpace" - the template brain to which the image that formed this MIP, was aligned. Typically, this is the JR2018 standard template brain from Bogovic et al. 2018.

"channel" - number of the channel from the aligned image stack that is represented by this MIP.

"mountingProtocol" - the protocol used to prepare brain sample for imaging.

"matchingPixels" - the number of overlapping pixels between query (searched.id) and taregt (nb.id).

"gradientAreaGap " - unsure, seeking clarification from NeuronBridge

"normalizedGapScore" - unsure, seeking clarification from NeuronBridge

"normalizedScore" - the matching score, created by examining the overlapped pixel number and color depth. If the color and xy position of the pixel match between the mask and the searching data, then the approach here will count it as a positive matching score

"searched.id" - the nb.id you searched with, i.e. given to the function call

"nb.id" - the 'NeuronBridge ID' for the MIP file.

See also

Examples

nb.info.em = neuronbridge_info("542634818")

#> Warning: invalid factor level, NA generated

nb.hits = neuronbridge_hits(nb.info.em$nb.id[1])

#> Warning: invalid factor level, NA generated

# View(nb.hits[1:10,]) # see the top 10 hits

# One can also do this in one with:
nb.search = neuronbridge_search("542634818")

#> Warning: invalid factor level, NA generated

#> Warning: invalid factor level, NA generated

## However, this differs from the above in that every MIP
## associated with the given id, is searched. For a connectome
## neuron this is just 1, but for a GAL4 line with MCFO data it
## can be many quite different images.

# Note that nb.info.em here is actually an attribute you can see
nb.info.em = attr(nb.search,"search")