Open Fiji on your computer and use it to convert your .nrrd files into color MIPs

Open Fiji on your computer and use it to convert your .nrrd files into color MIPs for searches. You must have downloaded Fiji onto your machine and also have followed the instructions here (search tool) and here (MIP generator) to install the Fiji plugins from Scientific Computing at Janelia Research Campus, for the Color MIP Search. When running the MIP generator, select directory that contains 3D stacks to be color depth MIP then select directory for saving MIPs. Pay attention to the parameters you can then adjust. Skeleton bolded MIP can be useful for data coming from skeleton or even mesh sources.

nrrd_to_mip(fiji.path = NULL, macro = NULL, MinMem = MaxMem, MaxMem = "2500m")

Arguments

fiji.path

path to your system's Fiji. If NULL, attempts to find using getOption("jimpipeline.fiji").

macro

path to the Fiji macro used to create the MIP file. Defaults to the macro contained within this R package: Color_Depth_MIP_batch_0308_2021.ijm.

MinMem

lower memory limit

MaxMem

upper memory limit

Value

a MIP files, saved as .tif, converted from a folder of selected .nrrd files

See also

Examples

if (FALSE) { # \dontrun{
nrrd_to_mip()
} # }